update i686 screenshot
the previous screenshot was taken on a interly different setup so
as you can see there is slight resources usage increase but a bit
more accurate in terms of what people face with mainstream
distributions. on Ubuntu it uses ~210MB of memory while idle, on
Arch Linux ~160MB so do not blame me if your distro it trying to
conquer the world.